Add a Makefile and more documentation for the demo
This adds a simple `make demo` target that builds the demo with
`gcc`. The readme is also updated to mention the make target and give
an example of how to run it. The example should make clearer what is
to be passed in as the `dbname` parameter; initially I tried `db.zip`
and passing in the `out_v1` directory path before realizing that each
of the database files is separate and that the full path should be
given.