Add stddef.h for size_t in header file
Got an error that size_t was not defined. I am no expert, but I suppose
this is the way to fix it.

Probably not an issue in most cases, but I ran into this, so someone
else might and then they will not have to think about it. I assume folks
usually will have this defined as a happy coincidence before including
the header.